+++ packages/core/src/connectivity/scp-cf/destination/destination.spec.ts
@@ -10,7 +10,7 @@ import {
 } from './destination';
 import { Destination } from './destination-service-types';
 
-describe('Destination edmx-parser', () => {
+describe('parseDestination', () => {
   it('destination is parsed correctly ', () => {
     const actual = parseDestination(basicMultipleResponse[0]);
 
@@ -28,7 +28,7 @@ describe('Destination edmx-parser', () => {
     expect(actual).toMatchObject(expected);
   });
 
-  it('TrustAll is correctly parsed', () => {
+  it("'TrustAll' is correctly parsed", () => {
     const actual = parseDestination({
       ...basicMultipleResponse[0],
       TrustAll: 'TRUE'
@@ -36,15 +36,96 @@ describe('Destination edmx-parser', () => {
     expect(actual.isTrustingAllCertificates).toBe(true);
   });
 
-  it('isTrustingAllCertificates is set to false when TrustAll is not "TRUE"', () => {
+  it('`isTrustingAllCertificates` is set to false when TrustAll is not "TRUE"', () => {
     const actual = parseDestination({
       ...basicMultipleResponse[0],
-      TrustAll: 'TRUEE'
+      TrustAll: 'NOT_TRUE'
     });
     expect(actual.isTrustingAllCertificates).toBe(false);
   });
 
-  it('isTrustingAllCertificates is set as boolean', () => {
+  it('parses certificates', () => {
+    const actual = parseDestination(certificateSingleResponse);
+    expect(actual.authentication).toBe('ClientCertificateAuthentication');
+    expect(actual.keyStoreName).toBe('key.p12');
+    expect(actual.keyStorePassword).toBe('password');
+    expect(actual.certificates!.length).toBe(1);
+    expect(actual.certificates![0]).toEqual({
+      name: 'key.p12',
+      content: 'base64string',
+      type: 'CERTIFICATE'
+    });
+  });
+
+  it('initializes certificates as empty array if none are present', () => {
+    const actual = parseDestination(certificateMultipleResponse[0]);
+    expect(actual.authentication).toBe('ClientCertificateAuthentication');
+    expect(actual.keyStoreName).toBe('key.p12');
+    expect(actual.keyStorePassword).toBe('password');
+    expect(actual.certificates!.length).toBe(0);
+  });
+
+  it('parses additional headers and query parameters', () => {
+    const destination = parseDestination({
+      URL: '',
+      'URL.headers.additionalHeader1': 'additionalHeader1',
+      'URL.queries.additionalQueryParam': 'additionalQueryParam',
+      'URL.headers.additionalHeader2': 'additionalHeader2'
+    });
+    expect(destination).toEqual(
+      expect.objectContaining({
+        headers: {
+          additionalHeader1: 'additionalHeader1',
+          additionalHeader2: 'additionalHeader2'
+        },
+        queryParameters: {
+          additionalQueryParam: 'additionalQueryParam'
+        }
+      })
+    );
+  });
+
+  it('parses additional headers and query parameters as undefined if not provided', () => {
+    const destination = parseDestination({
+      URL: ''
+    });
+
+    expect(destination).not.toHaveProperty('headers');
+    expect(destination).not.toHaveProperty('queryParameters');
+  });
+
+  it('throws an error if there is no `URL` given', () => {
+    expect(() =>
+      parseDestination({
+        Name: 'DEST'
+      } as any)
+    ).toThrowErrorMatchingInlineSnapshot(
+      '"Property \'URL\' of destination configuration must not be undefined."'
+    );
+  });
+
+  it("does not throw when there is no `URL` for destinations with type other than 'HTTP' or undefined", () => {
+    expect(() =>
+      parseDestination({ Type: 'RFC' } as DestinationConfiguration)
+    ).not.toThrow();
+  });
+});
+
+describe('sanitizeDestination', () => {
+  it('should be idempotent', () => {
+    const destination = {
+      username: 'username',
+      password: 'password',
+      name: 'DEST',
+      url: 'https://example.com'
+    };
+
+    const oneTime = sanitizeDestination(destination);
+    const twoTimes = sanitizeDestination(oneTime);
+    expect(oneTime).toMatchObject(twoTimes);
+  });
+
+  it('sets `isTrustingAllCertificates` as boolean', () => {
     const destination = {
       name: 'destination',
       isTrustingAllCertificates: true,
@@ -61,65 +142,19 @@ describe('Destination edmx-parser', () => {
     expect(actual).toMatchObject(expected);
   });
 
-  it('certificates are parsed correctly', () => {
-    const actual = parseDestination(certificateSingleResponse);
-    expect(actual.authentication).toBe('ClientCertificateAuthentication');
-    expect(actual.keyStoreName).toBe('key.p12');
-    expect(actual.keyStorePassword).toBe('password');
-    expect(actual.certificates!.length).toBe(1);
-    expect(actual.certificates![0]).toEqual({
-      name: 'key.p12',
-      content: 'base64string',
-      type: 'CERTIFICATE'
-    });
-  });
-
-  it('certificates are initialized as empty array if none are present', () => {
-    const actual = parseDestination(certificateMultipleResponse[0]);
-    expect(actual.authentication).toBe('ClientCertificateAuthentication');
-    expect(actual.keyStoreName).toBe('key.p12');
-    expect(actual.keyStorePassword).toBe('password');
-    expect(actual.certificates!.length).toBe(0);
-  });
-
-  it('sanitizeDestination should be idempotent', () => {
-    const destination = {
-      username: 'username',
-      password: 'password',
-      name: 'DEST',
-      url: 'https://example.com'
-    };
-
-    const oneTime = sanitizeDestination(destination);
-    const twoTimes = sanitizeDestination(oneTime);
-    expect(oneTime).toMatchObject(twoTimes);
-  });
-
-  it('parseDestination throws an error if there is no URL given', () => {
-    expect(() =>
-      parseDestination({
-        Name: 'DEST'
-      } as any)
-    ).toThrowErrorMatchingSnapshot();
-  });
-
-  it('sanitizeDestination throws an error if there is no url given', () => {
+  it('throws an error if there is no `url` given', () => {
     expect(() =>
       sanitizeDestination({
         username: 'username',
         password: 'password',
         name: 'DEST'
       })
-    ).toThrowErrorMatchingSnapshot();
+    ).toThrowErrorMatchingInlineSnapshot(
+      '"Property \'url\' of destination input must not be undefined."'
+    );
   });
 
-  it("parseDestination does not throw when there is not url for destinations with type other than 'HTTP' or undefined", () => {
-    expect(() =>
-      parseDestination({ Type: 'RFC' } as DestinationConfiguration)
-    ).not.toThrow();
-  });
-
-  it("sanitizeDestination does not throw when there is not url for destinations with type other than 'HTTP' or undefined", () => {
+  it("does not throw when there is no `url` for destinations with type other than 'HTTP' or undefined", () => {
     expect(() => sanitizeDestination({ type: 'RFC' })).not.toThrow();
   });
 });