+++ packages/core/src/connectivity/scp-cf/destination/destination.ts
@@ -55,7 +55,68 @@ export function parseDestination(
     }
   );
 
-  return sanitizeDestination(destination);
+  const additionalHeadersAndQueryParameters =
+    getAdditionalHeadersAndQueryParameters(destinationConfig);
+
+  return sanitizeDestination({
+    ...destination,
+    ...additionalHeadersAndQueryParameters
+  });
+}
+
+/**
+ * Get either additional headers or query parameters from a destination, depending on the given prefix.
+ * @param destinationConfig Original destination config that could include additional headers or query parameters.
+ * @param originalKeyPrefix This is what the additional header and query keys start with, when specified in the original destination config.
+ * @returns An object with either the headers or query parameters and their values, depending on the `originalKeyPrefix`.
+ */
+function getAdditionalProperties(
+  destinationConfig: DestinationConfiguration,
+  originalKeyPrefix: 'URL.headers.' | 'URL.queries.'
+): Record<string, any> {
+  const relevantConfigEntries = Object.entries(destinationConfig).filter(
+    ([key]) => key.startsWith(originalKeyPrefix)
+  );
+
+  return relevantConfigEntries.reduce(
+    (additionalProperties: Record<string, any>, [originalKey, value]) => {
+      const headerKey = originalKey.replace(originalKeyPrefix, '');
+      additionalProperties[headerKey] = value;
+      return additionalProperties;
+    },
+    {}
+  );
+}
+
+/**
+ * Get additional headers and/or query parameters from a destination.
+ * Destinations can specify additional headers and/or query parameters, that should be added to every request against the given destination.
+ * They are specified in the following format:
+ * `URL.headers.<header-name>` or `URL.queries.<query-parameter-name>`
+ * @param destinationConfig Original destination config that could include additional headers or query parameters.
+ * @returns An object with either the headers or query parameters and their values, depending on the `originalKeyPrefix`.
+ */
+function getAdditionalHeadersAndQueryParameters(
+  destinationConfig: DestinationConfiguration
+): Pick<Destination, 'headers' | 'queryParameters'> {
+  const additionalHeaders = getAdditionalProperties(
+    destinationConfig,
+    'URL.headers.'
+  );
+  const additionalQueryParameters = getAdditionalProperties(
+    destinationConfig,
+    'URL.queries.'
+  );
+
+  const additionalProperties = {};
+  if (Object.keys(additionalHeaders).length) {
+    additionalProperties['headers'] = additionalHeaders;
+  }
+  if (Object.keys(additionalQueryParameters).length) {
+    additionalProperties['queryParameters'] = additionalQueryParameters;
+  }
+
+  return additionalProperties;
 }
 
 function getDestinationConfig(